requirements.txt
fastapi
uvicorn[standard]
gunicorn
sqlalchemy
pandas
requests
python-dotenv
pydantic
